+++ test/test_checks.py
@@ -28,6 +28,8 @@ class CommaTestChecks(TestCase):
         comma_checker = CommaChecker(None, filename=get_absolute_path('data/bad_list.py'))
         self.assertEqual(list(comma_checker.get_comma_errors(comma_checker.get_file_contents())), [
             {'col': 5, 'line': 4, 'message': 'C812 missing trailing comma'},
+            {'col': 5, 'line': 10, 'message': 'C812 missing trailing comma'},
+            {'col': 5, 'line': 17, 'message': 'C812 missing trailing comma'},
         ])
 
     def test_bad_function_call(self):
@@ -98,6 +100,11 @@ class CommaTestChecks(TestCase):
 
     def test_comma_not_required_in_parenth_form_string_splits(self):
         fixture = 'data/multiline_string.py'
+        comma_checker = CommaChecker(None, filename=get_absolute_path(fixture))
+        self.assertEqual(list(comma_checker.get_comma_errors(comma_checker.get_file_contents())), [])
+
+    def test_comma_not_required_in_comment_lines(self):
+        fixture = 'data/good_list.py'
         comma_checker = CommaChecker(None, filename=get_absolute_path(fixture))
         self.assertEqual(list(comma_checker.get_comma_errors(comma_checker.get_file_contents())), [])